fix: request-client should support conversion and network checks
